# docs/v1/swagger.yaml
basePath: /api/web-api-gateway/v1
definitions:
  domain.CompanyUpdate:
    properties:
      address:
        type: string
      country:
        type: string
      email:
        type: string
      is_active:
        type: boolean
      is_approved:
        type: boolean
      name_en:
        type: string
      name_ru:
        type: string
      phone:
        type: string
      timezone:
        type: string
      website:
        type: string
    type: object
  domain.CreateCompany:
    properties:
      address:
        example: г. Алматы
        type: string
      country:
        example: KZ
        type: string
      email:
        example: example@example.com
        type: string
      is_active:
        example: true
        type: boolean
      is_approved:
        example: true
        type: boolean
      name_en:
        example: OOO ROGA I COPUTA
        type: string
      name_ru:
        example: ООО Рога и копыта
        type: string
      phone:
        example: "+77777777777"
        type: string
      timezone:
        example: Asia/Almaty
        type: string
      website:
        example: www.rogakopyta.kz
        type: string
    required:
    - address
    - email
    - name_ru
    - phone
    type: object
...
  domain.ErrorResponse:
    properties:
      code:
        type: integer
      data: {}
      is_error:
        type: boolean
      message:
        type: string
    type: object
...
  domain.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  domain.SuccessResponse:
    properties:
      data: {}
      is_error:
        type: boolean
      total_count:
        type: integer
    type: object
...
host: 91.243.71.100:8080
info:
  contact:
    email: ru.system.ru@gmail.com
    name: ru.system.ru@gmail.com
  description: API gateway
  title: Web api gateway API
  version: "1.0"
paths:
...
  /company/:
    get:
      consumes:
      - application/json
      description: |-
        Получение информации по всем компаниям.
        Только super admin может получать информацию.
      operationId: get-companies
      parameters:
      - description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: sort
        required: true
        type: string
      - default: name_ru
        description: Field to sort by
        enum:
        - id
        - name_ru
        - name_en
        - country
        - address
        - phone
        - email
        - website
        - is_active
        - created_at
        - updated_at
        - is_approved
        in: query
        name: sort_field
        required: true
        type: string
      - description: limit query param
        in: query
        name: limit
        required: true
        type: integer
      - description: offset query param
        in: query
        name: offset
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get companies
      tags:
      - company
    post:
      consumes:
      - application/json
      description: |-
        Создание компании.
        Только super admin может создавать компании
      operationId: create-company
      parameters:
      - description: Company info
        in: body
        name: company
        required: true
        schema:
          $ref: '#/definitions/domain.CreateCompany'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create company
      tags:
      - company
  /company/{id}:
    delete:
      consumes:
      - application/json
      description: |-
        Удаление компании.
        Только super admin может удалять компании.
      operationId: delete-company
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete company
      tags:
      - company
    get:
      consumes:
      - application/json
      description: |-
        Получение информации о компании по ID.
        Только super admin может получать информацию по id компании.
      operationId: get-company
      parameters:
      - description: Company ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get company
      tags:
      - company
    put:
      consumes:
      - application/json
      description: |-
        Обновление компании.
        Только super admin может обновлять active & approve компании
        Для обновления указывать только необходимые поля.
      operationId: update-company
      parameters:
      - description: Company ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      - description: Company info
        in: body
        name: company
        required: true
        schema:
          $ref: '#/definitions/domain.CompanyUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update company
      tags:
      - company
  /company/info:
    get:
      consumes:
      - application/json
      description: Получение информации о компании для пользователя
      operationId: get-company-info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get company info
      tags:
      - company
  /geo/city/list:
    get:
      consumes:
      - application/json
      description: Список городов
      operationId: get-city-list
      parameters:
      - description: country code query param
        example: RU
        in: query
        name: country_code
        required: true
        type: string
      - description: admin code query param
        example: "63"
        in: query
        name: admin_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get city list
      tags:
      - geo
  /geo/country/list:
    get:
      consumes:
      - application/json
      description: Список стран
      operationId: get-country-list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get country list
      tags:
      - geo
  /geo/region/list:
    get:
      consumes:
      - application/json
      description: Список регионов/областей
      operationId: get-region-list
      parameters:
      - description: country code query param
        example: RU
        in: query
        name: country_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get region list
      tags:
      - geo
...
  /user/company:
    get:
      consumes:
      - application/json
      description: Получение списка пользователей компании.
      operationId: get-users
      parameters:
      - description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: sort
        required: true
        type: string
      - default: name
        description: Field to sort by
        enum:
        - id
        - username
        - name
        - email
        - phone
        - created_at
        - updated_at
        - last_login
        - is_active
        - role
        - country
        - is_approved
        - position
        in: query
        name: sort_field
        required: true
        type: string
      - description: limit query param
        in: query
        name: limit
        required: true
        type: integer
      - description: offset query param
        in: query
        name: offset
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get users
      tags:
      - user
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"